refactor: divided sub components into folders with their own scss files
