+++ app/(pages)/about-us/_components/Archive/_carousel/carousel.module.scss
@@ -7,254 +7,4 @@
     --slide-size: 0%;
     display: flex;
     flex-direction:column;
-
-    .carousel__text_section {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        padding-left: 12%;
-        color: #123041;
-        font-family: "Metropolis";
-        font-size: 3rem;
-        font-style: normal;
-        font-weight: 700;
-        line-height: normal;
-        letter-spacing: 0.96px;
-        width: 90vw;
-
-        .text_content {
-            color: #123041;
-            font-family: "Metropolis";
-            font-size: 3rem;
-            font-style: normal;
-            font-weight: 700;
-            line-height: normal;
-            letter-spacing: 0.96px;
-            text-wrap: wrap;
-
-            @media (max-width: 768px) {
-                font-size: 2rem;
-                width: 311px;
-                height: 56px;
-            }
-        }
-
-        .controls {
-            display: grid;
-            grid-template-columns: auto 1fr;
-            justify-content: space-between;
-            gap: 1.2rem;
-
-            @media (max-width: 768px) {
-                position: absolute;
-                margin-left: 0%;
-                margin-top: 40vw;
-                display: none;
-            }
-
-            .buttons {
-                display: flex;
-                gap: 30px;
-
-                @media (max-width: 768px) {
-                    gap: 0px;
-                }
-
-                .single__button {
-                    -webkit-tap-highlight-color: rgba(var(--text-high-contrast-rgb-value), 0.5);
-                    -webkit-appearance: none;
-                    appearance: none;
-                    background-color: transparent;
-                    touch-action: manipulation;
-                    display: inline-flex;
-                    text-decoration: none;
-                    cursor: pointer;
-                    border: 0;
-                    padding: 0;
-                    margin: 0;
-                    box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
-                    width: 3.6rem;
-                    height: 3.6rem;
-                    z-index: 1;
-                    border-radius: 50%;
-                    color: var(--text-body);
-                    display: flex;
-                    align-items: center;
-                    justify-content: center;
-                    background-color: #88D8DD;
-
-                    @media (max-width: 768px) {
-                        width: 34px;
-                        height: 34px;
-                    }
-
-                    .single__button__svg {
-                        width: 35%;
-                        height: 35%;
-                    }
-                }
-
-
-                .single__button:disabled {
-                    background-color: #DBDBDB;
-                    color:grey;
-                }
-            }
-        }
-    }
-
-    .carousel__viewport {
-        padding-left: 12%;
-        padding-top: 10%;
-        padding-bottom: 60px;
-        width: 90%;
-
-        .carousel__container {
-            backface-visibility: hidden;
-            display: flex;
-            touch-action: pan-y;
-
-            .carousel__slide {
-                flex: 0 0 var(--slide-size);
-                min-width: 10;
-                padding-left: 0;
-            }
-        }
-    }
-
-
-    .carousel__progress_section {
-        display: flex;
-        justify-content: center;
-
-
-        @media (max-width: 768px) {
-            gap: 20px;
-        }
-
-        .prev_button_mobile {
-            display: none;
-            
-            @media (max-width: 768px) {
-                display: block ;
-            }
-
-            .single__button {
-                -webkit-tap-highlight-color: rgba(var(--text-high-contrast-rgb-value), 0.5);
-                -webkit-appearance: none;
-                appearance: none;
-                background-color: transparent;
-                touch-action: manipulation;
-                display: inline-flex;
-                text-decoration: none;
-                cursor: pointer;
-                border: 0;
-                padding: 0;
-                margin: 0;
-                box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
-                width: 3.6rem;
-                height: 3.6rem;
-                z-index: 1;
-                border-radius: 50%;
-                color: var(--text-body);
-                display: flex;
-                align-items: center;
-                justify-content: center;
-                background-color: #88D8DD;
-
-                @media (max-width: 768px) {
-                    width: 34px;
-                    height: 34px;
-                }
-
-                .single__button__svg {
-                    width: 35%;
-                    height: 35%;
-                }
-            }
-
-            .single__button:disabled {
-                background-color: #DBDBDB;
-                color:grey;
-            }
-        }
-
-        .carousel__progress {
-            border-radius: 1.8rem;
-            box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
-            background-color: #B3B3B3;;
-            position: relative;
-            height: 0.6rem;
-            justify-self: flex-end;
-            align-self: center;
-            width: 75vw;
-
-            overflow: hidden;
-
-            @media (max-width: 768px) {
-                width: 60vw;
-                max-width: 1080px;
-            }
-
-            .carousel__progress__bar {
-                background-color: #005271;;
-                position: absolute;
-                width: 17%;
-                border-radius: 10px;
-                top: 0;
-                bottom: 0;
-                left: 0%;
-
-                @media (max-width: 768px) {
-                    width: 10%;
-                }
-            }
-        }
-
-        .next_button_mobile {
-            display: none;
-
-            @media (max-width: 768px) {
-                display: block ;
-            }
-
-            .single__button {
-                -webkit-tap-highlight-color: rgba(var(--text-high-contrast-rgb-value), 0.5);
-                -webkit-appearance: none;
-                appearance: none;
-                background-color: transparent;
-                touch-action: manipulation;
-                display: inline-flex;
-                text-decoration: none;
-                cursor: pointer;
-                border: 0;
-                padding: 0;
-                margin: 0;
-                box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
-                width: 3.6rem;
-                height: 3.6rem;
-                z-index: 1;
-                border-radius: 50%;
-                color: var(--text-body);
-                display: flex;
-                align-items: center;
-                justify-content: center;
-                background-color: #88D8DD;
-
-                @media (max-width: 768px) {
-                    width: 34px;
-                    height: 34px;
-                }
-
-                .single__button__svg {
-                    width: 35%;
-                    height: 35%;
-                }
-            }
-            .single__button:disabled {
-                background-color: #DBDBDB;
-                color:grey;
-            }
-        }
-    }
   }
