+++ app/(pages)/about-us/_components/Archive/_carousel/components/CarouselArrowButtons/CarouselArrowButtons.module.scss
@@ -1,0 +1,38 @@
+.single__button {
+                -webkit-tap-highlight-color: rgba(var(--text-high-contrast-rgb-value), 0.5);
+                -webkit-appearance: none;
+                appearance: none;
+                background-color: transparent;
+                touch-action: manipulation;
+                display: inline-flex;
+                text-decoration: none;
+                cursor: pointer;
+                border: 0;
+                padding: 0;
+                margin: 0;
+                box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
+                width: 3.6rem;
+                height: 3.6rem;
+                z-index: 1;
+                border-radius: 50%;
+                color: var(--text-body);
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                background-color: #88D8DD;
+
+                @media (max-width: 768px) {
+                    width: 34px;
+                    height: 34px;
+                }
+
+                .single__button__svg {
+                    width: 35%;
+                    height: 35%;
+                }
+            }
+
+            .single__button:disabled {
+                background-color: #DBDBDB;
+                color:grey;
+            }
