+++ app/(pages)/about-us/_components/Archive/_carousel/components/CarouselContent/CarouselContent.module.scss
@@ -1,0 +1,18 @@
+.carousel__viewport {
+    padding-left: 12%;
+    padding-top: 10%;
+    padding-bottom: 60px;
+    width: 90%;
+
+    .carousel__container {
+        backface-visibility: hidden;
+        display: flex;
+        touch-action: pan-y;
+
+        .carousel__slide {
+            flex: 0 0 var(--slide-size);
+            min-width: 10;
+            padding-left: 0;
+        }
+    }
+}
