+++ app/(pages)/about-us/_components/Archive/_carousel/components/CarouselProgress/CarouselProgress.module.scss
@@ -1,0 +1,57 @@
+.carousel__progress_section {
+    display: flex;
+    justify-content: center;
+
+
+    @media (max-width: 768px) {
+        gap: 20px;
+    }
+
+    .prev_button_mobile {
+        display: none;
+        
+        @media (max-width: 768px) {
+            display: block ;
+        }
+    }
+
+    .carousel__progress {
+        border-radius: 1.8rem;
+        box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
+        background-color: #B3B3B3;;
+        position: relative;
+        height: 0.6rem;
+        justify-self: flex-end;
+        align-self: center;
+        width: 75vw;
+
+        overflow: hidden;
+
+        @media (max-width: 768px) {
+            width: 60vw;
+            max-width: 1080px;
+        }
+
+        .carousel__progress__bar {
+            background-color: #005271;;
+            position: absolute;
+            width: 17%;
+            border-radius: 10px;
+            top: 0;
+            bottom: 0;
+            left: 0%;
+
+            @media (max-width: 768px) {
+                width: 10%;
+            }
+        }
+    }
+
+    .next_button_mobile {
+        display: none;
+
+        @media (max-width: 768px) {
+            display: block ;
+        }
+    }
+}
